packages/tools/calendar-tool/pyproject.toml
[project]
name = "jarvis-calendar"
version = "0.1.0"
description = "Google Calendar integration tool for Jarvis"
dependencies = [
    "jarvis-shared",
    "google-auth>=2.29.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.140.0",
]
requires-python = ">=3.11"

[tool.uv.sources]
jarvis-shared = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
